pyproject.toml
[tool.poetry]
name = "RAGPilot"
version = "0.1.0"
description = ""
authors = ["nick <nickchen1998@gmail.com>"]
readme = "README.md"
packages = [{include = "RAGPilot"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"
selenium = "^4.31.0"
fake-useragent = "^2.2.0"
openai = "^1.76.0"
langchain-openai = "^0.3.14"
django-rest-framework = "^0.1.0"
celery = "^5.5.2"
redis = "^5.0.0"
pgvector = "^0.4.1"
langchain-cohere = "^0.4.4"
pandas = "^2.3.0"
pymongo = "^4.13.1"
gunicorn = "^21.2.0"
django-allauth = "^65.9.0"
langgraph = "^0.4.9"
pypdf = "^5.6.1"
langchain-experimental = "^0.3.4"
whitenoise = "^6.9.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
